[IMP] project_forecast: allow more action on gantt dialog
This commit allow to do more action on forecasts without leaving the gantt view. Such as deleting repeated forecasts (and stopping the recurrency), deleting future forecats (and stopping the recurrency), sending a single one per mail, or save & sending one of them at once.

This had to be done in js as there is no way to put object buttons in the footer of dialog. And forecast edition on gantt use a FormViewDialog. Plus we don't want the buttons to appear on the standard form.

related task #1929079