[IMP] helpdesk_timesheet: time reinvoicing option
Before this commit, when "use timesheet" was checked on a team, the project
and task fields were always visible, which can bring confusion to user. Indded
it was not clear that selecting a task can make the timesheet of a ticket
reinvoice through a Sales Order (with the SO line of the linked task).

This commit aims to only display the task (and project) fields when the new
"reinvoicing time" options is checked on the helpdesk team. We now considere that
only tracking time, there is no need to display the project on the itcket, as we
track time for internal reporting. All time spend should then be in the same "service
container" (project/analytic account). When checking the new option, we might want to
set the time spend on the ticket in a specific analytic account (one per customer
for instance), and so we need to change the project per ticket.

Task-1969885

closes odoo/enterprise#4344

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>