[IMP] ocn_client: send FCM Push Notifications for "System" messages
Before this commit, users only received FCM Push Notifications from
a record if someone sent a message or was mentioned in a log note.
-> only mail.message of type "comment".

After this commit if a user follow a record it will be able to receive
a notification on his mobile phone for each subtypes depending on
the model.
-> all mail.message with type "comment" or "system".

Example with project.task:
If a user has subscribed to "Activities" and "Stage Changed" subtypes
for "project1" record, he will now receive notifications when someone
creates a new activity on this record and will be notified if the stage
of this record changed.

Note that a user will receive these notifications only if "handle
notification in Odoo" is checked in their preferences and if they
enabled the given subtype notification on the given record.

Task ID : 1943362