[REF][IMP] project_forecast: revamp and introduce recurrency
Makes it easier to visualize data and allows more precision and control over the forecasts spans.

The two main fields of forecast formerly date_start and date_end were switched to datetime fields. Tis allows us more precise forecasts (like fr switches in a restaurant. The fields were renamed to start_datetime and end_datetime as well for clarity

We also implement a few features to help with forecast generation.
The user will probably create forecasts that are very similar on some
interval. Like each day of the week, or every 13th of the month.

Since creating every one forecast by hand will probably be cumbersome,
we added tools to automate a bit this creation process.

Recurrency will repeat the creation of a forecast based on some interval
and until a certain date, (every forecast created will be offseted from previous by this interval).
Also, this will allow to run a cron that will continue repeating for
some time. Avoiding to create too much objects at once. Thus the new
model.

ALso, since gantt view opens a dialog and a form view within, it was
covenient to include the repeat options in the form of a forecast. This was done through 2 wizards: the repeat wizard is used to repeat an already existing forecast. While the create wizard has a mix of field from the forecast form and the repeat wizard, allow to create and repeat a forecast at once.

Duplicate takes every forecast visible on the screen that fit in the
period (larger forecasts will be ignored). And recreates them in the
next period. repeated forecasts are ignored.
In order to achieve this we had to create a new js-class inheriting
gantt view (as of now, adding a button in the control panel is not
supported from xml or python).

Related to task #1929079