[FIX] web_studio: report editor cards color
Since rev. odoo/odoo@ddf2d2d the card-body has a default color.
As the Studio report editor sidebar is all grey/black, we do not want
this default color. A first fix (see odoo/enterprise@3627585) has been
done to correctly change the color on the card-body but the sidebar DOM
was not correct. Some components (section in the cards) did not have the
`card-body` class.

All the components now have a correct DOM, with the `card-body` being
set on their parent.

Task 1978603

closes odoo/enterprise#4248

Signed-off-by: VincentSchippefilt <VincentSchippefilt@users.noreply.github.com>