[IMP] hr_payroll: Move schedule_pay from contract to structure
Schedule pay depends on the salary structure, not the contract.
e.g. for the same contract, the regular pay is every month,
but the 13th month is every year.
Hence the field is moved to `hr.payroll.structure` and made
a related field on `hr.contract`.

Note: the field schedule_pay is currently unused.
It could probably be removed completly remove on the contract.