[IMP] documents: adds dynamic colors to tag categories
This commit adds the many2many display of tags on the documents' kanban
card.

To allow easy visual mapping of tags with their categories, this commit
also adds a dynamic coloring of tag categories based on their current
display order that is displayed both in the search panel and on the
kanban card.

Task: #1974652

closes odoo/enterprise#4489

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>