[IMP] hr_payroll: Improve work entries generation flow
Purpose
=======
Ease the payroll flow. Avoid as many clicks as possible.

Specification
=============
Generation / Conflicts
----------------------
- Work entries are automatically generated for past and current months.
- Work entries are also generated for new contracts. e.g. work entries have already
been generated but a new contract is open in the middle of the month => generate new work entries for this contract.
- If a work entry is manually created, conflicts should automatically be checked
- If a work entry is manually edited (date start and date stop), conflicts should automatically be checked
- A work entry without a work entry type is marked as conflicting.
Adding a type should remove the conflict, removing a type should make the conflict.
- If a work entry of type leave is completly outside the contract's calendar,
it is suspicious and it should be marked as conflicting.

Leaves
------
- If there is a non approved leave at the same time as some work entries, the work entries
are in conflict.
- If the leave is approved:
- A work entry corresponding to the leave is created.
- Existing work entries completety included in the leave are deleted
- Existing work entries that are not completely included are marked as conflicting.
- If the leave is refused, existing work entries are kept and marked as not conflicting.
- Conflicting work entries can be manually validated without approving/refusing the leave.
- If a leave is refused although it was previously validated (=> a corresponding work entry
was created), the corresponding work entry should disappear and "attendance" work entries
should be re-created instead.

Display
-------
In gantt and calendar views:
- Display only one button: "Generate Payslips"
- The button is disabled while there are conflicts. If there are no conflicts, payslips can
be directly generated.
- If there are conflicts, display a warning: "We found some conflicts, click here to revise them" which
leads to only conflicting work entries.

Note:
Conflict conditions are more and more complex/refined but the reason
is not necesarily precisly reported to end users.
The way errors are managed and later how the warning is displayed
should probably be improved.

closes odoo/enterprise#4316

Signed-off-by: Romain Libert (rli) <rli@odoo.com>