[FIX] web_studio: allow to drop images in report editor
Since rev. odoo/enterprise@27664e7 it was not possible anymore to drop
an image in the report editor. Basically nothing happened when you
dropped the image. The instruction that opened the modal had been
wrongly removed in the adaptation to native promises.

Note that the test has been adapted because the `onload` attribute is
not bound synchronously anymore and cannot thus be called. This is not
an issue because calling `onload` is not needed anymore.

closes odoo/enterprise#4299

Signed-off-by: Christophe Simonis <chs@odoo.com>