[IMP] web_map: add map view
The view uses the leaflets library to display the map.
Documentation link: https://leafletjs.com/reference-1.4.0.html.

This view was created to display records on a map. The records are displayed using markers.
The view also allows to display informations using popups. Each of the record displayed must
have a many2one to res.partner since coordinates and address are stored
on the res.partner model. The view is also able to show the routing between the records

2 apis are used to provide the tiles and the geoforwarding (conversion from address
to coordinates)

The first api is openstreetmap it does not requires any token.
The second is mapbox, it requires a token, check out this link to create
an account: https://account.mapbox.com/auth/signup/

The only api that provides the routing feature is mapbox

This view was created to display records on a map. The records are displayed using markers.
The view also allows to display informations using popups. Each of the record displayed must have a many2one to res.partner since the coordinates
and address are stored on the res.partner model

The root element of the map_view is <map>
The root's element is able to take multiple attributes:
- res_partner: should contain the res.partner many2one.
	if not provided, the view will create a map that don't show records
- default_order: overrides the model's default order. The new order will be determided
	by the provided field.
	The field provided must be part of the model on which the view is implemented
	not from res.partner
- routing: determine if the routing between markers will be computed of shown.
	For the route to be computed there must be more than 2 located records
	(i.e: the record is linked to a partner and the partner has either an address
	or valid coordinates). A valid MapBox token must also be provided.
The <marker-popup> element defines what fields are displayed inside the marker's popup
	-For each field that has to appear in the popup a field tag <field> must be created
	The tag takes 2 attributes
		-name which is the name of the field
		-string which is the title of the field. it'll be displayed before the
		field's content

taskid: 1931206