[IMP] ocn_client: associate Android channel for each kind of notification
This commit allows us to use Android channel.
For example a user is now able to change de priority of a specific
channel in the mobile app or disable it.

We distinguish 4 types of notifications:
- Direct Message
- Channel Message
- Following
- At mention

If none of theses channels is specified, a default one (created
in the mobile app) will be used.

If someone mentions you on a following record, "At mention" will be
used instead of "Following" because we consider that this type
of priority is more important. The user must receive it even
if he disabled "Following" channel in his Android device.

Because the payload may be different we can't send all theses
notification at the same time anymore.
We had to split into chunks depending if it's a mention or not.

Note that Android channels only work starting from Android 8.0 (Oreo)
See: https://developer.android.com/training/notify-user/channels

Task ID : 1848517