[IMP] sale_timesheet: improve demo data
previously sale_timesheet created orphans task. Another commit in
community links them to a project and adds stages to that project and
it's tasks. Problem is that this project's only task was being linked to
demo forecasts. Since it was the only one task the search for it was too
simple to keep working when more were added. Added a search criteria
based on the timesheets linked to that task

related task #1985386

closes odoo/enterprise#4315

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>