[IMP] mrp_workorder: workcenter kanban by default
Use the primary button 'work orders' will open the kanban view.
However the button that use a filter as 'to launch', 'in progress', ...
will open the list view. It should always open the kanban view since the
view is designer for operator and it should be easily clickable.

This commit overrides the action that open the list view and
replace with the kanban view.
view.

Task - 1967745

closes odoo/enterprise#4151

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>