[FIX] web_gantt: proper form view parameter passing
In gantt view, we can open a form view by clicking on a cell.
The form  view used should be the one from the action, or specified in
the context. But currently the "default" form view is always the one used.

This commit fixes the issue by properly sending the form view parameter
to the FormViewDialog created to contain the view.

Task-1943792