[IMP] hr_contract_salary: Make applicant's certificate required
Purpose
=======
Currently, on the salary configurator frontend, the certificate
is a radio button, with a default value selected.
Some applicants don't change the default value, even if it
is not their certificate.
Applicant should always explicitely set their certificate.

Specification
=============
Make the certificate a required field without a default value.
To reuse the existing validation logic of the salary JS, the input
is now a `select` instead of a `radio`.