[IMP] web_studio: remove 'Set another view' option from studio
Currently in the studio, we have an option 'Set another view' on the activated
views, using which we can select / change the view to be displayed from the
available views related to current model.

However, this option is quite technical for standard users to understand.

This commit removes the option and related code. Also for those views which
do not contain any advanced option, the drop-down icon will be hidden to
avoid confusion.

task-1982032
closes #4381

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>