[FIX] hr_payroll: Send work days amounts to server
When creating/editing a payslip, an onchange computes
work days lines, including the amount associated with
each line.

However, the amount field is not in the view hence it
is not sent to the server when the record is saved.
This leads to having all work days amounts to 0.0
(see the payslip report).

Quick fix done here: add the amount field to the view.
This fix is not the cleanest fix, it fixes the symptom but not
the root of the problem: Business logic (generating work days lines)
should not be performed by an onchange.
The way work days lines are generated should be refactored.

closes odoo/enterprise#4461

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>