[IMP] project_forecast: added employee picture on gantt rows
Having a thumbnail of the employee on corresponding lines in the gantt view is useful for quickly identifying which employee we are talking about.

In order to do this, we have overrriden the Gantt's row's template to include some custom code that'll detect if this line is grouped by an employee (groupedBy === 'employee_id') and if so, retrieve it's id (resId). Then using a route to build the picture's url.

In the custom template, the t-name is very important as otherwise, the picture will be visible on every gantt view, when there's a group by employee.

related task #1929079