[IMP] web_studio: indicate undroppable zone while editing search view
Currently, when we drag a component in search view, it can not be dropped
everywhere. For example, a 'filter' component can not be dropped inside
'Group by' and 'Autocompletion Fields' section. But user can not see any
visual indication about which component can be dropped where.

This commit indicates 'undroppable' zone by adding 'text-muted' class to
such tables, as soon as the draggable component touches the view.

task-1978762

closes odoo/enterprise#4392

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>