[FIX] web_gantt: give context for RPC calls
Currently, the context was not propagated when the gantt view call
`read_group` or `gantt_unavailability`. This prevented the use of
context key dependant behavior in server method.

This commit add the propagation of the context in the rpc calls in
order to stay consistent with the rest of webclient.

Task-1943792