[FIX] account_reports: followup next action date in 'auto' mode
When making the followup of a partner without changing the next action date (and leaving it in 'auto' mode), the next action date auto was simply ignored and the partner didn't dissapear from the partners to process. This would eventually work, by luck, if for some reason the next action mode was set in 'manual' then reverted to 'auto'.

closes odoo/enterprise#4366

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>