[FIX] sign: prevent crash when searching more in many2one
In the Contract Templates form view, there was a crash when the
user clicked on 'Search More' in the many2one dropdown of fields
'sign_template_id' and 'contract_update_template_id' (both pointing
to model 'sign.template'). It was due to an override of the
ListController for that model, that assumed that the '$buttons'
property existed. As of saas-12.1, it's not the case anymore when
it is not necessary (whereas before, the $buttons were always
rendered, and sometimes not happened into the DOM), like in the
'Search more' dialog.

closes odoo/enterprise#4367

Signed-off-by: VincentSchippefilt <VincentSchippefilt@users.noreply.github.com>