[FIX] web_gantt: decoration with date fields
Currently on the gantt view we cannot use date fields as "decoration"
fields. Trying to do so brings a traceback.

An example use case is in Field Services where we want to display a
decoration to indicate when a timer is running on some task.

This commit aims to fix the issue by formatting those date fields in
_getDecorationEvalContext before they are sent to py.evaluate in
_evaluateDecoration so they can be evaluated correctly.

Task-1943792

closes odoo/enterprise#4372

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>