[FIX] account_reports, followup menu: filter on 'overdue invoices' status
Filtering partners on 'onverdue invoices' in the list view was previously including partners 'in need of action'. Although not entirely wrong (partners in need of action are a subset of partners with overdue invoies), it's more logical and more powerful to restrict the return of the search to partners having the status 'with overdue invocies'.

closes odoo/enterprise#4378

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>