[IMP] web_gantt: add groupedByField to pillsInfo
It might be useful for the row to know which field from the current model is used in the groupBy. Eg: if you want to display employee's picture the grupby might be by multiple fields, you want to know if the actual line refrs to an employee or another field.

related task #1929079