[IMP] web_studio: make bottom image clickable on kanban
Currently, in kanban record editor, the user image (image in kanban
bottom which might already exists or can be added using the studio hook)
is not selectalbe. So if we want to remove it from the view, it is not
possible.

The image is now selectable and can be deleted.

Task 1997542

closes odoo/enterprise#4412

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>