[FIX] *account_taxcloud: make creation of taxes work again
odoo/odoo@3936d655c4bfd845b6ead724cecdc8e7c62816eb introduced
account.tax.repartition.line. These are automatically created via a
default_get on account.tax.

Setting company_id explicitly on the tax when creating it causes it to
not be in the default_get vals, leading Odoo to try to create a
repartition line with company_id as None. This causes an error since
it's a required field.

Apply a fix inspired by the above commit where the company_id is set
in context instead.