[FIX] web_mobile: many2one_barcode not displayed anymore
Before this commit, the many2one_barcode field was working only in
debug mode. The problem was that the enterprise source code was
adding the widget in the registry before the community in normal mode
(this bug doesn't appeared in debug mode).

Now, with this commit, we explicitly wait for _field_registry and now
enterprise correctly override community.

My guess is after the update of jQuery3/Promise, the promise is always
asynchronous. So the code who add the widget in the registry is
executed sooner than before.

Similar bug in odoo/odoo@a54d5eede8604ba9601ff21727fffef826113879

closes odoo/enterprise#4394

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>