[IMP] industry_fsm_report: add task report feature
This commit brings the ability to print reports from tasks including
informations concerning the task itself, all the items on the sale
order created from the task and the fields in the task's worksheet.

Technically, this mecanism is very strange and out of Odoo scope: as
report to print are models generated on the fly (on
project.worksheet.template creation), we need to convert a form view
into a qweb template (attached to the worksheet template) so we can
render it as "custom" report part in task report to print.

FP request.

Task-1943792