[IMP] hr_payroll: Base payslip computation on work entries
Purpose
=======

In payslips, use generated work entries to compute work days lines, instead of
duplicating the values from the work entries to the resource calendar and
use the resource calendar to compute the payslips values:

1/ Avoid to do two times the same job
2/ Copying work entries to calendar brings timezone issues.
3/ This significantly reduces the code complexity

Specification
=============

Current behavior before PR:
Each time that work entry are modified, calendar is modified to stay consistent.
And when we want to generated a payslip, we use calendar to generate worked days lines.
So we make twice the same computation from calendar: once to generate work entries and once
to generate payslips.

Desired behavior after PR is merged:
We generate work entries and then we can use this work entries to generate payslips.
If we generate a payslip and no work entries was previously generated, it will automatically generate
work entries (with maximum future the end of the month).
On each contract, we save two dates: the generated_from and generated_to. Each time we want
to generate new work entries, we will check already generated part, and just generate missing interval.
If there is a gap in generated interval, it will be also generated.

TaskID: 2001623

closes odoo/enterprise#4501

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>