[IMP] web_studio: sync selected node with sidebar in report editor
When editing a report using studio, we have option to highlight the nodes
on editor with help of sidebar. When we hover on available accordian cards
in the sidebar, it's associcated node in the report is highlighted with a
red arrow. However, once we expand / select any parent node with sidebar,
the options related to that node are displayed, but on the report editor,
the initial node is stil selected and user can not get proper idea for
which node the properties are shown / expanded.

This commit fixes the issue by 'selecting' the same node on report
editor which is active (instead of initial one) on the sidebar.

task-1986834

closes odoo/enterprise#4407

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>