[IMP] industry_fsm_*: FSM task flow with states
Impacted modules: industry_fsm ,industry_fsm_report and sale_timesheet_enterprise

This commit split the "done" button action. A new one appears: "send report" to
send the worksheet to the customer. To do so, we needed a 'state' notion on task.
This commit introduces the flow and the different state the task should follow
during its lifetime:
- created in draft, the user can send report and/or create the SO
- "send report" will send the worksheet using the mail composer to use traditionnal
odoo tools. When sent, the task is marked in 'sent' state
- The "validate" action becomes primary. The modal to create the SO appears to user,
and the sales order can be generated. The task is marked as 'done' and moved to the
first 'is closed' stage found in the project

For usability purpose, the "validate" and "send report" buttons are hidden
when the employee is timsheeting (timer running).
In 'draft' and 'sent', the timer can be launched, but not when the SO is generated.

Task-1943792