[IMP] sale_subscription: automatic salesteam selection
This commit adds the automatic selection of the salesteam
for sale.subscription object so that changing the
salesperson will automatically changes the team if the new
salesperson has one. If they are not part of a team,
the current value is left unchanged.

This will avoid having the manually correct the salesteam
every time the salesperson is modified.

task-1984950 closes-#4417

closes odoo/enterprise#4417

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>