[FIX] sign: remove useless code crashing when using promises
To reproduce in 12.3: click on "Sign Directly" on an existing template with a
signature required from multiple people, fill the fields, and try to sign and
validate it.

Since https://github.com/odoo/enterprise/commit/5bded7e4ba6a35b3d655008933fc3787799ace68#diff-91972131dae188ced2f7ec3bfa9db27aR1172

The code is useless because it is writing back the same value that was just
assigned, without changing it in-between.

It didn't crash in 12.1 because before promises `this` was the deferred in that
situation, so it was possible to write on it, even if it wasn't useful.

Starting from 12.3, after we started to use promises, `this` is undefined in
that situation, so it crashes when we try to write on it.

Discussion in https://github.com/odoo/enterprise/pull/4384

closes odoo/enterprise#4409

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>