[IMP] hr_timesheet,helpdesk,project_forecast: Show remaining hours on tasks
Show the remaining hours to timesheet on tasks. Allows to quickly evaluate task's state.
This is done in the name_get of a task through a context key, because we need this different name on some views, not all.

This allows to know how much hours are left to timesheet on a task while forecasting without having to change context.

Task-1856139