[FIX] iot: Longpolling requests not stopped
We used to save both `event` and `action` requests in the listener.
When trying to stop the longpolling, if the last request that was sent
was an action, we would try to abort this one instead of the `event`.
This resulted in multiple `event` requests running at the same time.

To fix this, we save only the requests that are made on the
`hw_drivers/event` route in the listener. This way, we ensure that
the request that we kill when restarting the longpolling is the indeed
the polling and not an action.