[FIX] industry_fsm_report: prevent traceback upon creating a task from document
Previously when user tried to create a task from documents with the
module industry_fsm_report installed, it was throwing a traceback
because it was trying to access 'report_template_id' of project.project
which was of type None as we do not have project_id here.

This commit fixes this by modifying if condition to check for the valid
value of project_id first to move any further.

Task-2002519

closes odoo/enterprise#4443

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>