[IMP] *account_taxcloud: don't assume every error is related to settings
Whenever the TaxCloud API returned an error message Odoo would
mention:

"The configuration of TaxCloud is in the Accounting app, Settings
menu."

This is only useful if TaxCloud credentials were not set in
Settings. In case of another error (e.g. "the destination zip is
missing") this is confusing, the error has nothing to do with the
TaxCloud configuration in Settings.

Remove this message, we now have warnings that pop up before
confirming an SO or invoice so TaxCloud should be configured when
raising this error.