[IMP] various: update o_list_view selectors to o_list_table
The list view now has a new top-most div which was required for
the optional fields dropdown to work outside of the overflow-hidden
environement of the responsive table.

We decided to keep the more generic "o_list_view" class name for
the top-most div, so the table now has class "o_list_table" instead.