[IMP] account_taxcloud: advise to enable auto_apply on TaxCloud fp
Ideally this would be the default but it causes issues. It would break
a lot of tests since most demo customers are American and so all
SO/invoice confirmations hit TaxCloud and fail. Additionally it would
break manual testing by a user on runbot or demo for the same reason.

closes odoo/enterprise#3402

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>