[REF] web_enterprise: mail chatter topbar styles
The Chatter's topbar (containing the buttons like 'Send Message', 'Log
Note') was difficult to deal with: specially the positioning of its
content (a float into another float...).

So, this commit is a follow-up of the refactoring of the topbar's layout
using Flexbox. It simplifies and reduces the amount of (S)CSS needed to
achieve the same visual result.

Task ID: 1928573