[FIX] documents: Create documents with default res_id/res_model
Steps to reproduce:
1. Install documents and project
2. Go to res setting to enable saving project attachments to Documents
3. Edit a project
4. Click on the Document stat button
5. From there, create an ir.attachment
=> document.document is not created in the Document app.

To integrate an ir.attachment into the Documents app,
bridge modules implement the method `_create_document` to
create a document.document.

Currenly, when creating an ir.attachment, the create vals parameter
is passed to the method. It uses those vals to retrieve res_id
and res_model.
However, res_id and res_model might not be present in vals but
still be set at creation (e.g. it uses default values).
In this case the document.document cannot be created.
After this commit, res_id and res_model are correctly passed
to the `_create_document` method by using values from the
recordset instead of vals.