[FIX] voip: fix typo in c21f9ac54b23b57fd3a6193497a7aebeb3f923b7
This enumeration is meant to be done on `messages`, as indicated by the
name of the enum variables. There is a 1-to-1 mapping between `messages`
and `self` after calling super(), because each activity results in a new
message. Additionally, after super() the activities in `self` are
deleted and should not be used.

Related to odoo/enterprise#3441 and Task-1915004