[IMP] sign: template access flexbility
Before this commit, there were some inconsistencies in the way access
rights were handled on signature templates:
- managers could do everything
- users could only see document templates they had 'favourited'

This second option seems to make sense on paper, but it was actually
impossible to use from the UI: the field containing the 'users who
favourited' a template was entirely invisible; a manager could not
invite any users to a template; in practice this meant that 'Officers'
of
the app could only access templates they uploaded themselves (which is
not very useful -_-).

This commit somewhat changes this mechanism to make it more practical
and scalable:
- document templates can now be shared accross the whole company
	(an upcoming task is being considered to make this more
	granular, e.g based on Groups)
  => this only gives Read access on the template, allowing users to see
     it and create requests from it
- it is now possible to invite users to access a template on a
per-person basis using the existing (but previously) hidden many2many
field

A slight side-effect of this was that it was possible for users to see
documents they had not favourited in the kanban view; implying they
might
favourite it themselves from there (in case it was a public template).
Since public templates are only accessible for read operations, marking
it as a favourite raised an accesserror.
The view was adapted to prevent this error by hiding the star if a user
has no enough rights to write on templates which are not alreay in their
favourites.

Future changes to this app should include more granular controls and
display for templates, but this simple change will suffice for now.

    Co-authored-by: Victor Feyens <vfe@odoo.com>
    Co-authored-by: Damien Bouvy <dbo@odoo.com>