[IMP] iot, pos_iot: Reverse Six transactions
Add the possibility to reverse the last payment that has been done.

data.Reversal should be set in the response of a transaction to notify
that this transaction can be reversed.

We need to override `compute_extradue`. Before, if the due amount of
the last line was the same than the remaining due amount, the extradue
was not shown under the payment lines. Now, reversed transactions have
an amount of 0, which means that those two due amount will be equal.
We still need to show the extra due amount in this case.

TaskId: 1932416