[FIX] hr_payroll: Generate work entry for full days
Currently, `generate_work_entry` has a different behavior
if the type of parameters is `date` or `datetime`:
If the type is `date`, it implicitely set hours and minutes
to 00:00.
It is particularly confusing when `date_stop` is a `date`:
work entries are not generated for this day because it's
actually outside of the range.

This commit makes it consistent in all cases:
-hours and minutes are set to 00:00 for start date
-hours and minutes are set to 23:59 for end date

closes odoo/odoo#33906

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>

X-port/Cherry-pick of odoo/odoo@05ed9ebbe575aa4647eadd4897a3e0b15bb68f9f