[FIX] web_studio: correctly apply store attribute on new related field
Before this commit, when adding a new related field from the studio, it was
always stored by default regardless of the original field.

This commit fixes the issue by adding a proper 'store' attribute while adding
a related field. The attribute will be the same as the field which is selected
on the dialog. And thus it will make the behavior consistent / predictable.

task-1985710

closes odoo/enterprise#4484

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>