[FIX] sign: adapt to name and signature widget change
The name and signature widget needs to load additional xml
dependencies which are asynchronous so the start method of this
widget (originally called by 'replace' when the modal was opened)
cannot be called in the modal opened handler as it used to as
this breaks the loading promise chain since nothing would wait on
the promise returned by the start of the name and signature widget.

The start of this widget is now relegated to the willStart of the
dialog such that it can be waited for appropriately. However, the
start of name and signature used to call resetSignature, which would
not work properly unless the node is in the DOM, which is not yet
the case during willStart. So it was changed so that it would not
call resetSignature in its start method. It is now the responsibility
of the widget which uses name and signature to call resetSignature
when it is in the DOM, so the code of sign has to be adapted as well.

Closes odoo/enterprise#4537