[FIX] helpdesk_*: after sale actions access
Impacted modules: helpdesk_repair, helpdesk_sale and helpdesk_stock

If a user has no stock or sale access rights, he can't see the form view of
a ticket when the helpdesk team allows "return" and "repair".
This commit fixes that: the idea is a normal helpdesk user can see the count of
documents (repairs, returns, ...) but only create and see them if he has enough
access. When clicking on the stat button, he will have a warning message explaining
that he need more access to see the documents.

Task-2006909

closes odoo/enterprise#4592

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>