[IMP] mrp_workorder: improve workorder navigation in tablet view
Currently, the Work Order tablet flow can be improve in two ways:

- Firstly, the return button get back the WO views of all WO from the
workcenter. It's annoying when you come from a Manufacturing Order, so
now, this button will only come back a step in the crumbread.

- Secondly, when you done and close a WO from the tablet view, it could
be appreaciable to get the WO view for all other WO from the same MO
when you come from a Manufacturing Order.
So, now it's the case (it still display the WO for the same Work Center
when you come from the Work Center kanban view).

Task #1965962

closes odoo/enterprise#4877

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>